+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # HEAD
+- Constructive Solid Geometry [#10](https://github.com/stevelonny/jujutracer/pull/10)
 - Implement `Pigment` types and rudimental `BRDF` methods
 - Implement `Flat` renderer
 # Version 0.2.1